What was I smoking when I created this?

+++ filemanage.py
@@ -20,52 +20,27 @@ class File():
 		with open(self.path, "r", encoding=encoding) as f:
 			return f.readlines()
 	
-	def rewrite(self, content, encoding: str = "utf-8"):
-		"""Clears the file and writes to it"""
-		with open(self.path, "w+") as f:
+	def overwrite(self, content, encoding: str = "utf-8"):
+		"""Clears the file and writes to it."""
+		with open(self.path, "w+", encoding=encoding) as f:
 			f.write(content)
 		return File(self.path)
 	
 	def write(self, content, encoding: str = "utf-8"):
-		"""Does not clear the file and writes to it"""
-		with open(self.path, "w+") as f:
+		"""Does not clear the file and writes to it."""
+		with open(self.path, "a+", encoding=encoding) as f:
 			f.write(content)
 		return File(self.path)
 	
 	def create(self):
-		"""Creates a file. If the file exists it will be cleared"""
+		"""Creates a file. If the file exists it will be cleared."""
 		with open(self.path, "w+") as f:
 			f.close()
 		return File(self.path)
-
-# Everything below is meaningless, don't use it. Don't use filemanage at all
-def fw(path, content):
-	with open(path, "a+") as f:
-		f.write(content)
-
-def fwae(path, content):
-	with open(path, "w+") as f:
-		f.write(content)
-
-def fc(path):
-	with open(path, "w+") as f:
-		pass
-
-def fe(path):
-	with open(path, "w+") as f:
-		f.truncate()
-
-def fr(path):
-	with open(path, "r", encoding="utf-8") as f:
-		return f.read()
-
-def frl(path):
-	with open(path, "r", encoding="utf-8") as f:
-		return f.readlines()
-
-def getlen(file):
-    with open(file, "r+", encoding='utf-8') as f:
-        len = 0
-        for i in f:
-            len+=1
-        return len
+	
+	def clear(self):
+		"""Clears a file. If the file does not exist None is returned."""
+		if not os.path.isfile(self.path): return None
+		with open(self.path, "w") as f:
+			f.truncate()
+		return File(self.path)
